Project Refactoring
