--- ChatProcessor/ChatProcessor.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="CounterStrikeSharp.API" Version="1.0.318" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ChatProcessorApi\ChatProcessorApi.csproj" />
-  </ItemGroup>
-
-	<ItemGroup>
-		<None Update="lang\**\*.*" CopyToOutputDirectory="PreserveNewest" />
-	</ItemGroup>
-
-</Project>
