--- ChatProcessor/Utils/LangKey.cs
@@ -1,17 +1,0 @@
-﻿namespace ChatProcessor.Utils;
-
-public static class LangKey
-{
-    public const string CHAT_ALL = "Cstrike_Chat_All";
-    public const string CHAT_ALL_DEAD = "Cstrike_Chat_AllDead";
-
-    public const string CHAT_SPEC = "Cstrike_Chat_Spec";
-
-    public const string CHAT_T_DEAD = "Cstrike_Chat_T_Dead";
-    public const string CHAT_T = "Cstrike_Chat_T";
-    public const string CHAT_T_LOC = "Cstrike_Chat_T_Loc";
-
-    public const string CHAT_CT_DEAD = "Cstrike_Chat_CT_Dead";
-    public const string CHAT_CT = "Cstrike_Chat_CT";
-    public const string CHAT_CT_LOC = "Cstrike_Chat_CT_Loc";
-}
