+++ src/ChatProcessor.Core/Utils/LangKey.cs
@@ -1,0 +1,20 @@
+﻿namespace ChatProcessor.Utils;
+
+public static class LangKey
+{
+    public static class Chat
+    {
+        public const string All = "Cstrike_Chat_All";
+        public const string AllDead = "Cstrike_Chat_AllDead";
+
+        public const string Spec = "Cstrike_Chat_Spec";
+
+        public const string TDead = "Cstrike_Chat_T_Dead";
+        public const string T = "Cstrike_Chat_T";
+        public const string TLoc = "Cstrike_Chat_T_Loc";
+
+        public const string CtDead = "Cstrike_Chat_CT_Dead";
+        public const string Ct = "Cstrike_Chat_CT";
+        public const string CtLoc = "Cstrike_Chat_CT_Loc";   
+    }
+}
