+++ src/ChatProcessor.Core/lang/en.json
@@ -1,0 +1,285 @@
+﻿{
+	"Cstrike_Chat_CT_Loc": " {LightBlue}[СП] {0}{Green}﹫{2}{Default}: {1}",
+	"Cstrike_Chat_T_Loc": " {Yellow}[Т] {0}{Green}﹫{2}{Default}: {1}",
+	"Cstrike_Chat_CT_Dead": " {LightBlue}[СП] {0} {Grey}[DEAD]{Default}: {1}",
+	"Cstrike_Chat_T_Dead": " {Yellow}[Т] {0} {Grey}[DEAD]{Default}: {1}",
+	"Cstrike_Chat_CT": " {LightBlue}[СП] {0}{Default}: {1}",
+	"Cstrike_Chat_T": " {Yellow}[Т] {0}{Default}: {1}",
+	"Cstrike_Chat_All": " [ALL] {0}{Default}: {1}",
+	"Cstrike_Chat_AllDead": " [ALL] {0} {Grey}[DEAD]{Default}: {1}",
+	"Cstrike_Chat_AllSpec": " [ALL] {0} {Grey}[SPEC]{Default}: {1}",
+	"Cstrike_Chat_Spec": " {0} {Grey}[SPEC]{Default}: {1}",
+
+	"BombsiteA": "Bombsite A",
+	"BombsiteB": "Bombsite B",
+	"BombsiteC": "Bombsite C",
+	"Hostages": "Hostages",
+	"HostageRescueZone": "Hostage Rescue Zone",
+	"VipRescueZone": "VIP Rescue Zone",
+	"CTSpawn": "CT Start",
+	"TSpawn": "T Start",
+	"Bridge": "Bridge",
+	"Middle": "Middle",
+	"House": "House",
+	"Apartment": "Apartment",
+	"Apartments": "Apartments",
+	"Market": "Market",
+	"Sewers": "Sewers",
+	"Tunnel": "Tunnel",
+	"Ducts": "Ducts",
+	"Village": "Village",
+	"Roof": "Roof",
+	"Upstairs": "Upstairs",
+	"Downstairs": "Downstairs",
+	"Basement": "Basement",
+	"Crawlspace": "Crawlspace",
+	"Kitchen": "Kitchen",
+	"Inside": "Inside",
+	"Outside": "Outside",
+	"Tower": "Tower",
+	"WineCellar": "Wine Cellar",
+	"Garage": "Garage",
+	"Courtyard": "Courtyard",
+	"Water": "Water",
+	"FrontDoor": "Front Door",
+	"BackDoor": "Back Door",
+	"SideDoor": "Side Door",
+	"BackWay": "Back Way",
+	"FrontYard": "Front Yard",
+	"BackYard": "Back Yard",
+	"SideYard": "Side Yard",
+	"Lobby": "Lobby",
+	"Vault": "Vault",
+	"Elevator": "Elevator",
+	"DoubleDoors": "Double Doors",
+	"SecurityDoors": "Security Doors",
+	"LongHall": "Long Hall",
+	"SideHall": "Side Hall",
+	"FrontHall": "Front Hall",
+	"BackHall": "Back Hall",
+	"MainHall": "Main Hall",
+	"FarSide": "Far Side",
+	"Windows": "Windows",
+	"Window": "Window",
+	"Attic": "Attic",
+	"StorageRoom": "Storage Room",
+	"ProjectorRoom": "Projector Room",
+	"MeetingRoom": "Meeting Room",
+	"ConferenceRoom": "Conference Room",
+	"ComputerRoom": "Computer Room",
+	"BigOffice": "Big Office",
+	"LittleOffice": "Little Office",
+	"Dumpster": "Dumpster",
+	"Airplane": "Airplane",
+	"Underground": "Underground",
+	"Bunker": "Bunker",
+	"Mines": "Mines",
+	"Front": "Front",
+	"Back": "Back",
+	"Rear": "Rear",
+	"Side": "Side",
+	"Ramp": "Ramp",
+	"Underpass": "Underpass",
+	"Overpass": "Overpass",
+	"Stairs": "Stairs",
+	"Ladder": "Ladder",
+	"Gate": "Gate",
+	"GateHouse": "Gate House",
+	"LoadingDock": "Loading Dock",
+	"GuardHouse": "Guard House",
+	"Entrance": "Entrance",
+	"VendingMachines": "Vending Machines",
+	"Loft": "Loft",
+	"Balcony": "Balcony",
+	"Alley": "Alley",
+	"BackAlley": "Back Alley",
+	"SideAlley": "Side Alley",
+	"FrontRoom": "Front Room",
+	"BackRoom": "Back Room",
+	"SideRoom": "Side Room",
+	"Crates": "Crates",
+	"Truck": "Truck",
+	"Bedroom": "Bedroom",
+	"FamilyRoom": "Family Room",
+	"Bathroom": "Bathroom",
+	"LivingRoom": "Living Room",
+	"Den": "Den",
+	"Office": "Office",
+	"Atrium": "Atrium",
+	"Entryway": "Entryway",
+	"Foyer": "Foyer",
+	"Stairwell": "Stairwell",
+	"Fence": "Fence",
+	"Deck": "Deck",
+	"Porch": "Porch",
+	"Patio": "Patio",
+	"Wall": "Wall",
+	"BoatBar": "Boat Bar",
+	"BoatStorage": "Boat Storage",
+	"CTBar": "CT Bar",
+	"UpperCatwalks": "Upper Catwalks",
+	"LowerCatwalks": "Lower Catwalks",
+	"UpperCatwalk": "Upper Catwalk",
+	"UpperConveyor": "Upper Conveyor",
+	"GroundLevel": "Ground Level",
+	"LockerRoom": "Locker Room",
+	"BackEntrance": "Back Entrance",
+	"FrontEntrance": "Front Entrance",
+	"SnipersNest": "Sniper's Nest",
+	"BodyShop": "Body Shop",
+	"BankInterior": "Bank Interior",
+	"BombSite": "Bomb Site",
+	"BankExterior": "Bank Exterior",
+	"Street": "Street",
+	"GasStation": "Gas Station",
+	"SideEntrance": "Side Entrance",
+	"PalaceInterior": "Palace Interior",
+	"BackAlleys": "Back Alleys",
+	"CTSideUpper": "CT Side Upper",
+	"CTSideLower": "CT Side Lower",
+	"TSideUpper": "T Side Upper",
+	"TSideLower": "T Side Lower",
+	"UpperCarousel": "Upper Carousel",
+	"MidCarousel": "Mid Carousel",
+	"LowerCarousel": "Lower Carousel",
+	"CTCorridorDown": "CT Corridor Down",
+	"CTCorridorUp": "CT Corridor Up",
+	"TCorridorDown": "T Corridor Down",
+	"TCorridorUp": "T Corridor Up",
+	"LeftAlley": "Left Alley",
+	"RightAlley": "Right Alley",
+	"BackCourtyard": "Back Courtyard",
+	"FrontCourtyard": "Front Courtyard",
+	"MidArch": "Mid Arch",
+	"OutsideLong": "Outside Long",
+	"LongDoors": "Long Doors",
+	"AbovePit": "Above Pit",
+	"Pit": "Pit",
+	"APlatform": "A Platform",
+	"LongA": "Long A",
+	"ARamp": "A Ramp",
+	"BackofA": "Back of A",
+	"Bricks": "Bricks",
+	"ShortStairs": "Short Stairs",
+	"Short": "Short",
+	"Catwalk": "Catwalk",
+	"Mid": "Mid",
+	"TopofMid": "Top of Mid",
+	"MidDoors": "Mid Doors",
+	"ExtendedA": "Extended A",
+	"BDoors": "B Doors",
+	"Hole": "Hole",
+	"BPlatform": "B Platform",
+	"BackofB": "Back of B",
+	"UpperTunnel": "Upper Tunnel",
+	"TunnelStairs": "Tunnel Stairs",
+	"LowerTunnel": "Lower Tunnel",
+	"OutsideTunnel": "Outside Tunnel",
+	"UnderA": "Under A",
+	"TRamp": "T Ramp",
+	"HutRoof": "Hut Roof",
+	"Secret": "Secret",
+	"Crane": "Crane",
+	"Vents": "Vents",
+	"Decon": "Decon",
+	"Observation": "Observation",
+	"Admin": "Admin",
+	"Toxic": "Toxic",
+	"Mini": "Mini",
+	"Control": "Control Room",
+	"Heaven": "Heaven",
+	"Hell": "Hell",
+	"Vending": "Vending",
+	"Trophy": "Trophy",
+	"Silo": "Silo",
+	"SniperBox": "Sniper Box",
+	"Radio": "Radio",
+	"Hut": "Hut",
+	"Crows": "Crows",
+	"Squeaky": "Squeaky",
+	"Yellow": "Yellow",
+	"Main": "Main",
+	"CTRed": "CT Red",
+	"Red": "Red",
+	"Warehouse": "Warehouse",
+	"BackStairs": "Back Stairs",
+	"Rafters": "Rafters",
+	"Forklift": "Forklift",
+	"Garden": "Garden",
+	"Ruins": "Ruins",
+	"Banana": "Banana",
+	"Logs": "Logs",
+	"TStairs": "T Stairs",
+	"SecondMid": "Second Mid",
+	"Hay": "Hay",
+	"Quad": "Quad",
+	"Arch": "Arch",
+	"Library": "Library",
+	"Graveyard": "Graveyard",
+	"LowerMid": "Lower Mid",
+	"Floor50": "50th Floor",
+	"Floor51": "51st Floor",
+	"LadderBottom": "Ladder Bottom",
+	"LadderTop": "Ladder Top",
+	"River": "River",
+	"Shed": "Shed",
+	"Barn": "Barn",
+	"Cart": "Cart",
+	"PalaceAlley": "Palace Alley",
+	"PalaceTunnel": "Palace Tunnel",
+	"Shop": "Shop",
+	"Scaffolding": "Scaffolding",
+	"Connector": "Connector",
+	"TicketBooth": "Ticket Booth",
+	"Jungle": "Jungle",
+	"LowerPark": "Lower Park",
+	"UpperPark": "Upper Park",
+	"Fountain": "Fountain",
+	"Playground": "Playground",
+	"Restroom": "Restroom",
+	"Construction": "Construction",
+	"Canal": "Canal",
+	"Tunnels": "Tunnels",
+	"Pipe": "Pipe",
+	"Walkway": "Walkway",
+	"Rafts": "Rafts",
+	"MotorBoats": "Motorboats",
+	"SeaRock": "Sea Rock",
+	"BoatHouse": "Boathouse",
+	"SeaCliff": "Sea Cliff",
+	"Beach": "Beach",
+	"YellowHouse": "Yellow House",
+	"RedHouse": "Red House",
+	"Villa1stFloor": "Villa 1st Floor",
+	"Villa2ndFloor": "Villa 2nd Floor",
+	"DirtRoad": "Dirt Road",
+	"Road": "Road",
+	"BackRoad": "Back Road",
+	"SmallForest": "Small Forest",
+	"BigForest": "Big Forest",
+	"EastForest": "East Forest",
+	"WestForest": "West Forest",
+	"Glade": "Glade",
+	"DeadTrees": "Dead Trees",
+	"BarnRoof": "Barn Roof",
+	"PowerLines": "Power Lines",
+	"Helicopter": "Helicopter",
+	"Helicopters": "Helicopters",
+	"SidePath": "Side Path",
+	"Field": "Field",
+	"InsertionA": "Insertion A",
+	"InsertionB": "Insertion B",
+	"InsertionC": "Insertion C",
+	"InsertionD": "Insertion D",
+	"DriveWay": "Driveway",
+	"Ivy": "Ivy",
+	"PopDog": "Popdog",
+	"TMain": "T Main",
+	"Tunnel1": "Tunnel 1",
+	"Tunnel2": "Tunnel 2",
+	"ElectricalBox": "Electrical Box",
+	"Gravelpit": "Gravel Pit",
+	"Forest": "Forest",
+	"Tents": "Tents",
+	"Dock": "Dock"
+}
