+++ src/Modules/ChatProcessor.LocalGag/ChatProcessorLocalGag.cs
