+++ src/Modules/ChatProcessor.LocalGag/lang/en.json
