+++ src/Modules/ChatProcessor.Tags/ChatProcessor.Tags.csproj
@@ -11,7 +11,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\ChatProcessorApi\ChatProcessorApi.csproj" />
+		<ProjectReference Include="..\..\ChatProcessor.Api\ChatProcessor.Api.csproj" />
 	</ItemGroup>
 
 </Project>
