+++ src/Modules/ChatProcessor.Tags/Config/Tag.cs
