+++ src/Modules/ChatProcessor.Tags/Config/TagsConfig.cs
